builder/requirements.txt
# Required Python packages get listed here, one per line.
# Reccomended to lock the version number to avoid unexpected changes.

# torchvision==0.14.1+cu117

runpod==1.3.0
torchvision
huggingface-hub==0.17.3
bitsandbytes==0.41.1
transformers==4.33.3
xformers==0.0.22
accelerate==0.23.0
scipy == 1.11.2

git+https://github.com/huggingface/diffusers.git@e731ae0ec817649bf2c15f9f719269d57062696c
